+++ CHANGELOG.md
@@ -2,6 +2,54 @@
 
 ## [Unreleased]
 
+## [1.15.0](../../compare/1.14.0..1.15.0) - 2023-01-11
+### Added
+- `intersects()` function to `html` modile.
+
+## [1.14.0](../../compare/1.13.0..1.14.0) - 2023-01-11
+### Added
+- `selector()` function to `html` module.
+
+## [1.13.0](../../compare/1.12.0..1.13.0) - 2022-12-29
+### Added
+- `sleep()` function.
+
+## [1.12.0](../../compare/1.11.0..1.12.0) - 2022-12-25
+### Added
+- Functions: `object.deepMerge()`, `util.isPrimitive()`
+
+### Changed
+- Function names: `file.getBase()` -> `file.base()`, `file.getExt()` -> `file.ext()`
+- Test names are more descriptive
+
+## [1.11.0](../../compare/1.10.0..1.11.0) - 2022-12-18
+### Added
+- `array` module with functions: `diff()` and `uniq()`
+- Sourcemaps
+
+## [1.10.0](../../compare/1.9.0..1.10.0) - 2022-12-15
+### Added
+- `time` module with function `ms()`
+
+## [1.9.0](../../compare/1.8.0..1.9.0) - 2022-12-12
+### Added
+- `file` module with functions: `getBase()`, `getExt()`
+
+## [1.8.0](../../compare/1.7.0..1.8.0) - 2022-12-09
+### Added
+- Functions to `object` module: `equal()`, `partlyEqual()`, `strictlyEqual()` and `clone()`
+
+## [1.7.0](../../compare/1.6.0..1.7.0) - 2022-12-09
+### Added
+- `object` module with functions: `deepSeal()` and `deepFreeze()`
+
+## [1.6.0](../../compare/1.5.0..1.6.0) - 2022-12-02
+### Added
+- `roman` module with functions: `parse()`, `stringify()` and `valid()`
+
+### Removed
+- Main `index.js` entry file
+
 ## [1.5.0](../../compare/1.4.2..1.5.0) - 2022-11-23
 ### Added
 - Main `index.js` file that exports all available modules